+++ apps/web/package.json
@@ -17,6 +17,7 @@
     "@mantine/hooks": "^8.1.1",
     "@mantine/notifications": "^8.1.3",
     "@next/third-parties": "^15.3.5",
+    "@repo/common": "workspace:*",
     "@supabase/supabase-js": "^2.50.1",
     "@tailwindcss/postcss": "^4.1.10",
     "@vis.gl/react-google-maps": "^1.5.3",
