+++ apps/web/src/entities/user.ts
@@ -1,15 +1,4 @@
 import { z } from 'zod';
-import type { Profile } from './profile';
-import type { UserArtistFollow } from './userArtistFollow';
-
-export type User = {
-  id: string;
-  email: string;
-  createdAt: Date;
-  updatedAt: Date;
-  profile: Profile;
-  followingArtists: UserArtistFollow[];
-};
 
 export const signupUserSchema = z.object({
   email: z
