+++ apps/web/src/features/artist/components/EventSelectButton.tsx
@@ -1,7 +1,7 @@
 import { Badge, Image } from '@mantine/core';
+import type { Event } from '@repo/common';
 import { FaCheck, FaMapMarkerAlt } from 'react-icons/fa';
 import { DEFAULT_IMAGE_URL, getAreaLabel } from '@/constants';
-import type { Event } from '@/entities/event';
 import dayjs from '@/libs/dayjs';
 
 type Props = {
