+++ apps/web/src/features/event/components/ArtistSelectButton.tsx
@@ -1,7 +1,7 @@
 import { Avatar } from '@mantine/core';
+import type { Artist } from '@repo/common';
 import { FaCheck } from 'react-icons/fa';
 import { DEFAULT_IMAGE_URL } from '@/constants';
-import type { Artist } from '@/entities/artist';
 
 type Props = {
   artist: Artist;
