+++ apps/web/src/features/top/components/ArtistCard.tsx
@@ -1,8 +1,8 @@
 'use client';
 import { Card, Image } from '@mantine/core';
+import type { Artist } from '@repo/common';
 import Link from 'next/link';
 import { DEFAULT_IMAGE_URL } from '@/constants';
-import type { Artist } from '@/entities/artist';
 
 type Props = {
   artist: Artist;
