+++ apps/web/src/features/top/components/EventCard.tsx
@@ -1,8 +1,8 @@
 'use client';
 import { Card, Image } from '@mantine/core';
+import type { Event } from '@repo/common';
 import Link from 'next/link';
 import { DEFAULT_IMAGE_URL } from '@/constants';
-import type { Event } from '@/entities/event';
 import dayjs from '@/libs/dayjs';
 
 type Props = {
