+++ packages/common/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@repo/common",
+  "version": "1.0.0",
+  "description": "",
+  "main": "src/index.js",
+  "scripts": {
+    "lint:ts": "tsc --noEmit",
+    "lint:biome": "biome check ./src"
+  },
+  "devDependencies": {
+    "@repo/typescript-config": "workspace:*",
+    "typescript": "^5.6.3"
+  }
+}
